% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_within_shiny_server}
\alias{launch_within_shiny_server}
\title{Launch Application within shiny-server}
\usage{
launch_within_shiny_server()
}
\description{
Launch Application within shiny-server
}
